задание урока 4.2. Управление схемами баз данных - выполнила

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -82,7 +82,7 @@ public class StudentController {
     }
 
     @GetMapping("/averageAgeOfStudents")
-    public Integer getAverageAgeOfStudents() {
+    public Double getAverageAgeOfStudents() {
         return studentService.averageAgeOfStudents();
     }
